refactor: move isyntaxmigrater into spi
