+++ core/src/main/java/dk/gtz/graphedit/internal/DemoSyntaxFactory.java
@@ -6,8 +6,8 @@ import java.util.UUID;
 
 import dk.gtz.graphedit.model.ModelEdge;
 import dk.gtz.graphedit.model.ModelVertex;
-import dk.gtz.graphedit.model.migration.ISyntaxMigrater;
 import dk.gtz.graphedit.spi.ISyntaxFactory;
+import dk.gtz.graphedit.spi.ISyntaxMigrater;
 import dk.gtz.graphedit.tool.IToolbox;
 import dk.gtz.graphedit.view.EdgeController;
 import dk.gtz.graphedit.view.ModelEditorController;
