+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/lts/LTSSyntaxMigrater.java
@@ -8,9 +8,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import dk.gtz.graphedit.model.migration.ISyntaxMigrater;
 import dk.gtz.graphedit.plugins.syntaxes.lts.model.ModelState;
 import dk.gtz.graphedit.plugins.syntaxes.lts.model.ModelTransition;
+import dk.gtz.graphedit.spi.ISyntaxMigrater;
 import dk.gtz.graphedit.util.MetadataUtils;
 
 public class LTSSyntaxMigrater implements ISyntaxMigrater {
