+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/petrinet/PNSyntaxFactory.java
@@ -6,7 +6,6 @@ import java.util.UUID;
 
 import dk.gtz.graphedit.model.ModelEdge;
 import dk.gtz.graphedit.model.ModelVertex;
-import dk.gtz.graphedit.model.migration.ISyntaxMigrater;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.model.ModelArc;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.model.ModelPlace;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.model.ModelTransition;
@@ -19,6 +18,7 @@ import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelArc;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelPlace;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelTransition;
 import dk.gtz.graphedit.spi.ISyntaxFactory;
+import dk.gtz.graphedit.spi.ISyntaxMigrater;
 import dk.gtz.graphedit.tool.IToolbox;
 import dk.gtz.graphedit.tool.Toolbox;
 import dk.gtz.graphedit.view.ModelEditorController;
