+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/petrinet/PNSyntaxMigrater.java
@@ -8,10 +8,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import dk.gtz.graphedit.model.migration.ISyntaxMigrater;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.model.ModelArc;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.model.ModelPlace;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.model.ModelTransition;
+import dk.gtz.graphedit.spi.ISyntaxMigrater;
 import dk.gtz.graphedit.util.MetadataUtils;
 
 public class PNSyntaxMigrater implements ISyntaxMigrater {
